requirements.txt
addict==2.4.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
cssselect==1.2.0
discogs-wrapper-python==1.0.1
distro==1.9.0
Django==5.0.1
fake-useragent==1.4.0
idna==3.6
importlib-metadata==7.0.1
isort==5.13.2
lxml==5.1.0
maison==1.4.3
music-tag==0.4.3
mutagen==1.47.0
oauthlib==3.2.2
parse==1.20.1
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.6.0
pydantic_core==2.16.2
pyee==11.1.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
ruyaml==0.91.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tinytag==1.10.1
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
w3lib==2.1.2
websockets==12.0
yaml-indent==0.1.4
zipp==3.17.0
